Fixed DeviceMemory and Buffer lock, flush and invalidate functions. (#105)
DeviceMemory and Buffer lock, flush and invalidate functions now take uint64_t parameters, like Vulkan expects.
Added a way to compute an aligned size for Buffer<T> and VertexBuffer<T>, to prevent validation errors.
Fixed installation of plugins.

+++ source/Ashes/D3D11Renderer/Src/Miscellaneous/D3D11DeviceMemory.hpp
@@ -30,13 +30,13 @@ namespace d3d11_renderer
 				, ashes::MemoryRequirements const & requirements
 				, ashes::MemoryPropertyFlags flags );
 			virtual ~DeviceMemoryImpl() = default;
-			virtual uint8_t * lock( uint32_t offset
-				, uint32_t size
+			virtual uint8_t * lock( uint64_t offset
+				, uint64_t size
 				, ashes::MemoryMapFlags flags )const = 0;
-			virtual void flush( uint32_t offset
-				, uint32_t size )const = 0;
-			virtual void invalidate( uint32_t offset
-				, uint32_t size )const = 0;
+			virtual void flush( uint64_t offset
+				, uint64_t size )const = 0;
+			virtual void invalidate( uint64_t offset
+				, uint64_t size )const = 0;
 			virtual void unlock()const = 0;
 
 		protected:
@@ -81,19 +81,19 @@ namespace d3d11_renderer
 		/**
 		*\copydoc	ashes::DeviceMemory::lock
 		*/
-		uint8_t * lock( uint32_t offset
-			, uint32_t size
+		uint8_t * lock( uint64_t offset
+			, uint64_t size
 			, ashes::MemoryMapFlags flags )const override;
 		/**
 		*\copydoc	ashes::DeviceMemory::flush
 		*/
-		void flush( uint32_t offset
-			, uint32_t size )const override;
+		void flush( uint64_t offset
+			, uint64_t size )const override;
 		/**
 		*\copydoc	ashes::DeviceMemory::invalidate
 		*/
-		void invalidate( uint32_t offset
-			, uint32_t size )const override;
+		void invalidate( uint64_t offset
+			, uint64_t size )const override;
 		/**
 		*\copydoc	ashes::DeviceMemory::unlock
 		*/
