+++ source/Ashes/Gl3Renderer/Src/Miscellaneous/GlDeviceMemory.hpp
@@ -32,13 +32,13 @@ namespace gl_renderer
 				, GLuint boundResource
 				, GLuint boundTarget );
 			virtual ~DeviceMemoryImpl() = default;
-			virtual uint8_t * lock( uint32_t offset
-				, uint32_t size
+			virtual uint8_t * lock( uint64_t offset
+				, uint64_t size
 				, ashes::MemoryMapFlags flags )const = 0;
-			virtual void flush( uint32_t offset
-				, uint32_t size )const = 0;
-			virtual void invalidate( uint32_t offset
-				, uint32_t size )const = 0;
+			virtual void flush( uint64_t offset
+				, uint64_t size )const = 0;
+			virtual void invalidate( uint64_t offset
+				, uint64_t size )const = 0;
 			virtual void unlock()const = 0;
 
 		protected:
@@ -63,19 +63,19 @@ namespace gl_renderer
 		/**
 		*\copydoc	ashes::DeviceMemory::lock
 		*/
-		uint8_t * lock( uint32_t offset
-			, uint32_t size
+		uint8_t * lock( uint64_t offset
+			, uint64_t size
 			, ashes::MemoryMapFlags flags )const override;
 		/**
 		*\copydoc	ashes::DeviceMemory::flush
 		*/
-		void flush( uint32_t offset
-			, uint32_t size )const override;
+		void flush( uint64_t offset
+			, uint64_t size )const override;
 		/**
 		*\copydoc	ashes::DeviceMemory::invalidate
 		*/
-		void invalidate( uint32_t offset
-			, uint32_t size )const override;
+		void invalidate( uint64_t offset
+			, uint64_t size )const override;
 		/**
 		*\copydoc	ashes::DeviceMemory::unlock
 		*/
